Add WinForm and Presentation projects; update references
- Updated solution file to include "WinForm" and "BowlingMegabucks.TournamentManager.Presentation" projects.
- Added XML documentation to `IBowlerViewModel`, `RetrieveBowlerAdapter`, and `RetrieveBowlerViewModel`.
- Created `BowlingMegabucks.TournamentManager.Presentation.csproj` with necessary configurations.
- Updated project references in `BowlingMegabucks.TournamentManager.csproj` and `BowlingMegabucks.TournamentManager.UnitTests.csproj`.
- Added `PresentationExtensions.cs` for service collection extensions.
- Changed target framework in unit tests from `net9.0-windows` to `net9.0`.
