+++ source/Tournament Manager.sln
@@ -26,6 +26,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{8118061F
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BowlingMegabucks.TournamentManager.IntegrationTests", "tests\BowlingMegabucks.TournamentManager.IntegrationTests\BowlingMegabucks.TournamentManager.IntegrationTests.csproj", "{B401E683-9BAF-9BD0-B9D3-698A725E119E}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "WinForm", "WinForm", "{5D0361D9-264E-4288-BAC5-88B6AC4B3403}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BowlingMegabucks.TournamentManager.Presentation", "presentation\BowlingMegabucks.TournamentManager.Presentation\BowlingMegabucks.TournamentManager.Presentation.csproj", "{C84E56D2-9D46-4960-BE2E-3D20E1464E04}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -52,16 +56,22 @@ Global
 		{B401E683-9BAF-9BD0-B9D3-698A725E119E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B401E683-9BAF-9BD0-B9D3-698A725E119E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B401E683-9BAF-9BD0-B9D3-698A725E119E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C84E56D2-9D46-4960-BE2E-3D20E1464E04}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C84E56D2-9D46-4960-BE2E-3D20E1464E04}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C84E56D2-9D46-4960-BE2E-3D20E1464E04}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C84E56D2-9D46-4960-BE2E-3D20E1464E04}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{422D0156-DFC6-40A3-8013-176F92042B1D} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
+		{422D0156-DFC6-40A3-8013-176F92042B1D} = {5D0361D9-264E-4288-BAC5-88B6AC4B3403}
 		{8CD415EC-EB98-42F5-9667-2528A2DF4785} = {8118061F-7F40-4B67-8EE9-22ADA801568F}
 		{BA2BAB66-F25D-4788-A258-2F7B7405F5FE} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 		{1CBE8234-E2CB-43EB-979C-8E133EFDEC3E} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 		{B401E683-9BAF-9BD0-B9D3-698A725E119E} = {8118061F-7F40-4B67-8EE9-22ADA801568F}
+		{5D0361D9-264E-4288-BAC5-88B6AC4B3403} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
+		{C84E56D2-9D46-4960-BE2E-3D20E1464E04} = {5D0361D9-264E-4288-BAC5-88B6AC4B3403}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {67EEFFFE-F35B-4A04-BCED-3AB93F1E9B99}
