+++ source/main/BowlingMegabucks.TournamentManager.csproj
@@ -57,7 +57,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <ProjectReference Include="..\business logic\BowlingMegabucks.TournamentManager.BusinessLogic.csproj" />
+	  <ProjectReference Include="..\presentation\BowlingMegabucks.TournamentManager.Presentation\BowlingMegabucks.TournamentManager.Presentation.csproj" />
 	</ItemGroup>
 
 	<ItemGroup>
