+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/Bowlers/IBowlerViewModel.cs
@@ -3,36 +3,84 @@ using BowlingMegabucks.TournamentManager.Models;
 
 namespace BowlingMegabucks.TournamentManager.Bowlers;
 
-internal interface IViewModel
+/// <summary>
+/// 
+/// </summary>
+public interface IViewModel
 {
+    /// <summary>
+    /// 
+    /// </summary>
     BowlerId Id { get; }
 
+    /// <summary>
+    /// 
+    /// </summary>
     string FirstName { get; }
 
+    /// <summary>
+    /// 
+    /// </summary>
     string MiddleInitial { get; }
 
+    /// <summary>
+    /// 
+    /// </summary>
     string LastName { get; }
 
+    /// <summary>
+    /// 
+    /// </summary>
     string Suffix { get; }
 
+    /// <summary>
+    /// 
+    /// </summary>
     string StreetAddress { get; }
 
+    /// <summary>
+    /// 
+    /// </summary>
     string CityAddress { get; }
 
+    /// <summary>
+    /// 
+    /// </summary>
     string StateAddress { get; }
 
+    /// <summary>
+    /// 
+    /// </summary>
     string ZipCode { get; }
 
+    /// <summary>
+    /// 
+    /// </summary>
     string EmailAddress { get; }
 
+    /// <summary>
+    /// 
+    /// </summary>
     string PhoneNumber { get; }
 
+    /// <summary>
+    /// 
+    /// </summary>
     string USBCId { get; }
 
+    /// <summary>
+    /// 
+    /// </summary>
     DateOnly? DateOfBirth { get; }
 
+    /// <summary>
+    /// 
+    /// </summary>
     Models.Gender? Gender { get; }
 
+    /// <summary>
+    /// 
+    /// </summary>
     string SocialSecurityNumber { get; }
 }
